Can show custom dirs and read files under them
This commit has a first successful run of "SDDP", which means that
it is now only showing the folders it creates to represent SRR
numbers. It also means that users can read the files under those
folders because now the files contain their AWS information for
`sddp` so their location does not have to reflect what it is in
the bucket to be accurate.

NOTE: It seems there is still work to be done to fix the caching
code, as it seems to do a `ListObject` call and delete anything
in the system that does not abide by that call...