Fix issue with trying to insert duplicate files
I was stupidly looping with a variable instead of using the index
which is fine except I was sharing the pointer to the file name of
the variable, which meant the file system "seemed" to have multiple
files of the same name because the loop would overwrite the value
the pointer was pointing to... whoops. Now the loop doesn't overwrite
that name, meaning all the files can be seen.