Check http status codes on reads
Before, fusera was taking the happy path of assuming everything
is alright in the world and that APIs are nice and always give
you what you want. Sadly, this is not the case and fusera needed
to grow up and understand that bad things happen and all you can
do is map it to a file error and try to move on in life.