Refactor nr pkg to become sdl pkg
This refactor cleans up the API for interacting with the newly
named sdl pkg. It also changes fusera to work with a Signer
interface in order to abstract how fusera gets authorized access
to files. Obviously the sdl Client implements the Signer interface,
but it is a very simple interface to implement so that anyone can
now utilize the fuseralib for their own setup. All one would have to
do is pass fuseralib the appropriate parameters and a Signer that
provides signed urls when fusrea needs them.

Also, this commit adds an important bug fix which avoids mistaking long
lists of accessions as an accession list file.